+++ source/vm/frame.h
@@ -1,12 +1,14 @@
 #ifndef FRAME_H
 #define FRAME_H
 
+#include <cstdint>
+
 class Block;
 
 struct Frame {
   Frame() {}
-  Frame(int ip, Block *block) : ip(ip), block(block) {}
-  int ip;
+  Frame(uint64_t ip, Block *block) : ip(ip), block(block) {}
+  uint64_t ip;
   Block *block;
 };
